backend/requirements.txt
fastapi
uvicorn[standard]
websockets
liboqs-python
pycryptodome
sqlalchemy
python-dotenv
pytest
pytest-asyncio
httpx
